+++ R/tmapOnLoad.R
@@ -1,6 +1,6 @@
 # envir = environment()
 .onLoad = function(...) {
-	options(tmap.style = "white", tmap.mode = "plot", tmap.design.mode = FALSE,
+	options(tmap.style = "white", tmap.mode = "plot", mode_pool = NULL, tmap.design.mode = FALSE,
 			tmap.devel.mode = FALSE)
 	assign("defaultTmapOptions", .defaultTmapOptions, envir = .TMAP)
 	assign("tmapOptions", .defaultTmapOptions, envir = .TMAP)